feat: refactored voxelizer

+++ yolo-3d-poc/src/main/java/nicok/bac/yolo3d/off/Header.java
@@ -1,8 +1,8 @@
 package nicok.bac.yolo3d.off;
 
 public record Header(
-        long vertexCount,
-        long faceCount,
-        long edgeCount
+        int vertexCount,
+        int faceCount,
+        int edgeCount
 ) {
 }
